Refactored methods for handling erb and yml files out into separate modules
