+++ lib/hodor/environment.rb
@@ -1,20 +1,19 @@
 require 'singleton'
 
-require 'yaml'
-require 'erb'
 require 'log4r'
 require 'log4r/configurator'
 require 'tmpdir'
 require 'open4'
 require 'socket'
 require 'etc'
+require_relative 'util/yml_tools'
 
 include Log4r
 
 module Hodor
   class Environment
     include Singleton
-
+    include YmlTools
     attr_reader :logger
     attr_accessor :options
 
@@ -40,24 +39,6 @@ module Hodor
         puts "Error:  #{ex.message}"
       end if @logger.nil?
       @logger
-    end
-
-    def erb_sub(erb_body)
-      ERB.new(erb_body).result(self.instance_eval { binding })
-    end
-
-    def erb_load(filename, suppress_erb=false)
-      if File.exists?(filename)
-        file_contents = File.read(filename)
-        sub_content = suppress_erb ? file_contents : erb_sub(file_contents)
-        sub_content
-      elsif !filename.start_with?(root)
-        erb_load(File.join(root, filename))
-      end
-    end
-
-    def yml_load(filename) #, suppress_erb=false)
-      YAML.load(erb_load(filename, false)) # suppress_erb))
     end
 
     def terse?
@@ -242,47 +223,7 @@ module Hodor
       va << " -p #{settings[:ssh_port] || 22}" 
     end
 
-    def yml_expand(val, parents)
-      if val.is_a? String
-        val.gsub(/\$\{.+?\}/) { |match|
-          cv = match.split(/\${|}/)
-          expr = cv[1]
-          ups = expr.split('^')
-          parent_index = parents.length - ups.length
-          parent = parents[parent_index]
-          parent_key = ups[-1]
-          parent_key = parent_key[1..-1] if parent_key.start_with?(':')
-          if parent.has_key?(parent_key)
-            parent[parent_key]
-          elsif parent.has_key?(parent_key.to_sym)
-            parent[parent_key.to_sym]
-          else
-            parent_key
-          end
-        }
-      elsif val.is_a? Hash
-        more_parents = parents << val
-        val.each_pair do |k, v|
-          exp_val = yml_expand(v, more_parents)
-          val[k] = exp_val
-        end
-      else
-        val
-      end
-    end
 
-    def yml_flatten(parent_key, val)
-      flat_vals = []
-      if val.is_a? Hash
-        val.each_pair { |k, v|
-          flat_vals += yml_flatten("#{parent_key}.#{k}", v)
-        }
-      else
-        parent_key = parent_key[1..-1] if parent_key.start_with?('.')
-        flat_vals = ["#{parent_key} = #{val}"]
-      end
-      flat_vals
-    end
 
     # Run an ssh command, performing any optional variable expansion
     # on the command line that might be necessary.
